Moved 3 funcs to `event_filters.py`: get_highest_romantic_relation, check_relationship_value, get_personality_compatibility
