+++ scripts/events_module/relationship/romantic_events.py
@@ -14,11 +14,13 @@ from scripts.game_structure import constants
 from scripts.game_structure import game
 from scripts.game_structure.localization import load_lang_resource
 from scripts.utility import (
-    get_highest_romantic_relation,
     event_text_adjust,
-    get_personality_compatibility,
     process_text,
     change_relationship_values,
+)
+from scripts.events_module.event_filters import (
+    get_highest_romantic_relation,
+    get_personality_compatibility,
 )
